+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.9.1...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.9.2...HEAD) - TBD
+
+## [4.9.2](https://github.com/schemathesis/schemathesis/compare/v4.9.1...v4.9.2) - 2026-01-18
 
 ### :bug: Fixed